requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0

# Scientific computing
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.22.0

# Image processing
opencv-python>=4.8.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Progress bars
tqdm>=4.66.0

# Metrics and evaluation
lpips>=0.1.4

# Cell segmentation
cellpose>=2.3.0

# Utilities
pytorch-lightning>=2.0.0

# Development
pre-commit>=3.6.0
ruff>=0.1.9
